+++ Runtime/stdlib/python/simpleml/model/supervised/regression/_ridge.py
@@ -26,4 +26,4 @@ class RidgeRegression(Estimator):
         )
 
     def fit(self, train_data: DataType, labels: DataType, **kwargs) -> Model:
-        return RidgeRegressionModel(self._underlying.fit(train_data.toArray(), labels.toArray(), **kwargs))
+        return RidgeRegressionModel(self._underlying.fit(train_data.toArray(), labels.toArray().astype('float'), **kwargs))
